Enable collapsible sections of JSON
Sections of JSON code can now be hidden until they are expanded by the
user. To do this, just enclose the code to hide between two new lines,
one with SECTION_START and one with SECTION_END. As long as those words
are somewhere on those lines, it'll work.

To accomplish this I had to create a new Markdown renderer. Eventually
we may even extend this to do other things, but for now it just handles
this specific feature.