Add common "param" method for query parameters
Having a single common method for parameters in a table will let us
manage the formatting of those parameters in a saner way later on. As
it is, if we ever want to change how the type of a parameter is
displayed, we have to change it tediously across the entire docs.

We will have to do that once for this change but then we would never
have to do that again.