Switch JSON :minimal to :verbose
When using the print_json method, we almost never wanted to actually
print out *all* the fields of an object, so we were using the :minimal
option almost every time. Now :minimal has been replaced with :verbose,
its opposite. That way we have a sensible default, and we can only show
all fields when we really want to.

In the near future this may be expanded or revised to allow for those
extra fields to be displayed, but collapsed by default.